Chatbot process to grab info from weather API
Using the openweathermap API to get the weather within the city inputted by the user by injecting the city into the API's URL. Then parsing the JSON received from the API and outputting the information to the chat with a concatenated string. In the case of an error, returning a JSON which lets the user know that the weather grab was unsuccessful.